Ensure graphiql loads
- Fix manifest per: https://github.com/rmosolgo/graphiql-rails/issues/75
- Nest graphql_controller in Api module
- Clean up routes
