+++ config/routes.rb
@@ -1,10 +1,11 @@
 Rails.application.routes.draw do
   if Rails.env.development?
-    mount GraphiQL::Rails::Engine, at: "/graphiql", graphql_path: "/graphql"
+    mount GraphiQL::Rails::Engine, at: "/graphiql", graphql_path: "/api/graphql"
   end
-  post "/graphql", to: "graphql#execute"
+
   namespace :api do
     resources :tokens, only: :create
+    post "/graphql", to: "graphql#execute"
   end
 
   get "*path", to: 'client#index'
